+++ app/main/views/add_service.py
@@ -3,9 +3,10 @@ from flask_login import current_user
 from notifications_python_client.errors import HTTPError
 
 from app import billing_api_client, service_api_client
+from app.formatters import email_safe
 from app.main import main
 from app.main.forms import CreateNhsServiceForm, CreateServiceForm
-from app.utils import email_safe, user_is_gov_user, user_is_logged_in
+from app.utils import user_is_gov_user, user_is_logged_in
 
 
 def _create_service(service_name, organisation_type, email_from, form):
