+++ app/main/views/dashboard.py
@@ -20,11 +20,10 @@ from werkzeug.utils import redirect
 from app import (
     billing_api_client,
     current_service,
-    format_date_numeric,
-    format_datetime_numeric,
     service_api_client,
     template_statistics_client,
 )
+from app.formatters import format_date_numeric, format_datetime_numeric
 from app.main import main
 from app.statistics_utils import get_formatted_percentage
 from app.utils import (
