+++ tests/app/main/views/uploads/test_upload_contact_list.py
@@ -6,7 +6,7 @@ import pytest
 from flask import url_for
 from freezegun import freeze_time
 
-from app.utils import normalize_spaces
+from app.formatters import normalize_spaces
 from tests.conftest import SERVICE_ONE_ID
